Fix some comments
