+++ rust/parquet/src/reader/mod.rs
@@ -20,7 +20,6 @@
 //!
 //! This mod provides API for converting between arrow and parquet.
 
-pub mod page_iterator;
 pub mod schema;
 
 pub use self::schema::{parquet_to_arrow_schema, parquet_to_arrow_schema_by_columns};
